Issue #171: move test_build_dataset
